+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+- Added `--list-criteria` flag to `check` command for easier setup/debugging.
+
 ### Changed
 
 - Updated dependencies